Double travel time bins

+++ contribs/dvrp/src/main/java/org/matsim/contrib/dvrp/analysis/VehicleOccupancyProfileCalculator.java
@@ -182,7 +182,7 @@ public class VehicleOccupancyProfileCalculator
 		}
 		endTime = Math.min(endTime, analysisEndTime);
 
-		int timeInterval = timeDiscretizer.getTimeInterval();
+		double timeInterval = timeDiscretizer.getTimeInterval();
 		int fromIdx = timeDiscretizer.getIdx(beginTime);
 		int toIdx = timeDiscretizer.getIdx(endTime);
